blueprint import: unit tests
- Merging two similar unit tests files
